refactor a little
